Code review

+++ TASK-WeatherApp/SettingsScene/ViewModelSettingsScene/SettingsSceneViewModel.swift
@@ -11,13 +11,13 @@ import SnapKit
 
 class SettingsSceneViewModel {
     
-    var savedLocations = [CityWeather]()
+    var savedLocations = [WeatherInfo]()
     
     var coreDataService = CoreDataService.sharedInstance
     
     var userSettings = UserDefaultsService.fetchUpdated()
     
-    var refreshUISubject = PassthroughSubject<UserDefaultsService, Never>()
+    var refreshUISubject = CurrentValueSubject<Bool, Never>(true)
     
     init() {
         
@@ -25,27 +25,71 @@ class SettingsSceneViewModel {
             savedLocations = cities
         }
     }
+    
+    deinit {
+        print("SettingsSceneViewModel deinit")
+
+    }
 }
 
 extension SettingsSceneViewModel {
     
     func remove(at position: Int) {
         
-        let id = savedLocations[position].id
-        savedLocations.remove(at: position)
-        coreDataService.delete(id)
-        refreshUISubject.send(userSettings)
+        if let id = Int64(savedLocations[position].id) {
+            savedLocations.remove(at: position)
+            coreDataService.delete(id)
+            refreshUISubject.send(true)
+        }
     }
     
-    func saveUserSettings() {
+    func saveUserSettings(measurmentUnit: MeasurementUnits?, wantedCity position: Int?, conditions: [ConditionTypes]?) {
+        if let unitToSave = measurmentUnit {
+            
+            switch unitToSave {
+            case .imperial:
+                userSettings.measurmentUnit = .imperial
+                break
+            default:
+                userSettings.measurmentUnit = .metric
+                break
+            }
+        }
+        
+        if let index = position {
+            userSettings.lastCityId = savedLocations[index].id
+            UserDefaultsService.setShouldShowUserLocationWeather(false)
+        }
+        
+        userSettings.shouldShowHumidity = false
+        userSettings.shouldShowPressure = false
+        userSettings.shouldShowWindSpeed = false
+        
+        if let conditionsToSave = conditions {
+            for condition in conditionsToSave {
+                switch condition {
+                case .humidity:
+                    userSettings.shouldShowHumidity = true
+                    break
+                case .pressure:
+                    userSettings.shouldShowPressure = true
+                    break
+                case .windSpeed:
+                    userSettings.shouldShowWindSpeed = true
+                    break
+                }
+            }
+            
+        }
         
         UserDefaultsService.updateUserSettings(measurmentUnit: userSettings.measurmentUnit,
-                                               lastCityId: nil,
+                                               lastCityId: userSettings.lastCityId,
                                                shouldShowWindSpeed: userSettings.shouldShowWindSpeed,
                                                shouldShowPressure: userSettings.shouldShowPressure,
                                                shouldShowHumidity: userSettings.shouldShowHumidity)
         
         userSettings = UserDefaultsService.fetchUpdated()
         
+        refreshUISubject.send(true)
     }
 }
